refactor: complete Claude service extraction and clean up main.ts
- Replace all direct Claude instance management with service calls
- Add session management handlers using services
- Clean up main.ts by removing all Claude-specific logic
- Add reload/refresh handling to preserve sessions
- Fix TypeScript compilation error in code generation handler
- Main.ts now only contains IPC handler mappings to services
