+++ electron/main.js
@@ -3,8 +3,7 @@ import { fileURLToPath } from 'url';
 import { dirname, join } from 'path';
 import { spawn } from 'child_process';
 import Store from 'electron-store';
-import * as pty from 'node-pty';
-import { existsSync, unlinkSync, readdirSync, statSync } from 'fs';
+import { existsSync } from 'fs';
 import { readFile, mkdir } from 'fs/promises';
 import { watch as chokidarWatch } from 'chokidar';
 import { homedir } from 'os';
@@ -36,158 +35,6 @@ const fileWatchers = new Map();
 // Initialize Claude services
 const sessionService = new ClaudeSessionService(store);
 const instanceService = new ClaudeInstanceService(sessionService);
-// Helper function to start Claude with a specific session ID (used for retries)
-async function startClaudeWithSessionId(instanceId, workingDirectory, instanceName, runConfig, sessionId, attemptNumber = 0, totalAttempts = 1) {
-    try {
-        const claudeInfo = await ClaudeDetector.detectClaude(workingDirectory);
-        if (!claudeInfo) {
-            throw new Error('Claude CLI not found');
-        }
-        const debugArgs = process.env.CLAUDE_DEBUG === 'true' ? ['--debug'] : [];
-        // Build arguments for retry
-        let baseArgs;
-        if (sessionId) {
-            baseArgs = ['--resume', sessionId, ...debugArgs];
-            console.log(`[Retry ${attemptNumber}/${totalAttempts}] Attempting to resume with session ID: ${sessionId}`);
-        }
-        else {
-            // Generate new session ID
-            const newSessionId = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
-                const r = Math.random() * 16 | 0;
-                const v = c === 'x' ? r : (r & 0x3 | 0x8);
-                return v.toString(16);
-            });
-            baseArgs = ['--session-id', newSessionId, ...debugArgs];
-        }
-        let { command, args: commandArgs, useShell } = ClaudeDetector.getClaudeCommand(claudeInfo, baseArgs);
-        // Override with run config if provided
-        if (runConfig) {
-            if (runConfig.command) {
-                command = runConfig.command === 'claude' ? command : runConfig.command;
-            }
-            if (runConfig.args && runConfig.args.length > 0) {
-                const allArgs = sessionId
-                    ? ['--resume', sessionId, ...runConfig.args, ...debugArgs]
-                    : baseArgs.concat(runConfig.args);
-                const result = ClaudeDetector.getClaudeCommand(claudeInfo, allArgs);
-                command = result.command;
-                commandArgs = result.args;
-                useShell = result.useShell;
-            }
-        }
-        // Create the PTY instance
-        const claudePty = pty.spawn(command, commandArgs, {
-            name: 'xterm-color',
-            cols: 80,
-            rows: 30,
-            cwd: workingDirectory,
-            env: { ...process.env, TERM: 'xterm-256color' },
-            useConpty: false
-        });
-        // Store the instance
-        instanceService.setInstance(instanceId, claudePty);
-        // Get remaining fallback session IDs for further retries
-        const sessionOptions = sessionService.getSessionWithFallbacks(instanceId);
-        let localAttemptIndex = attemptNumber;
-        let isRetrying = false;
-        let isKillingForRetry = false;
-        let hasUpdatedSuccessfulSession = false; // Prevent duplicate session updates
-        // Setup data handler with retry detection
-        claudePty.onData((data) => {
-            // Send output to renderer
-            mainWindow?.webContents.send(`claude:output:${instanceId}`, data);
-            // Check for session restoration failure
-            if (!isRetrying && data.includes('No conversation found with session ID:')) {
-                console.log(`[Retry] Session ${sessionId} not found, checking for more fallbacks...`);
-                // Send failure status
-                mainWindow?.webContents.send(`claude:restoration-status:${instanceId}`, {
-                    status: 'failed',
-                    sessionId: sessionId,
-                    attemptNumber: localAttemptIndex + 1,
-                    totalAttempts: totalAttempts
-                });
-                // Check if we have more fallbacks
-                localAttemptIndex++;
-                if (localAttemptIndex < sessionOptions.fallbacks.length) {
-                    const nextSessionId = sessionOptions.fallbacks[localAttemptIndex];
-                    console.log(`[Retry] Attempting next fallback session ID: ${nextSessionId}`);
-                    // Send retry status
-                    mainWindow?.webContents.send(`claude:restoration-status:${instanceId}`, {
-                        status: 'retrying',
-                        sessionId: nextSessionId,
-                        attemptNumber: localAttemptIndex + 1,
-                        totalAttempts: totalAttempts
-                    });
-                    isRetrying = true;
-                    isKillingForRetry = true;
-                    // Kill current process
-                    try {
-                        claudePty.kill();
-                    }
-                    catch (error) {
-                        console.error('Error killing Claude process for retry:', error);
-                    }
-                    // Remove from instances
-                    instanceService.deleteInstance(instanceId);
-                    // Retry with next session ID
-                    setTimeout(async () => {
-                        const retryResult = await startClaudeWithSessionId(instanceId, workingDirectory, instanceName, runConfig, nextSessionId, localAttemptIndex, totalAttempts);
-                        if (!retryResult.success) {
-                            console.error(`Failed to retry with fallback session ID: ${retryResult.error}`);
-                            mainWindow?.webContents.send(`claude:restoration-status:${instanceId}`, {
-                                status: 'all-failed',
-                                error: retryResult.error,
-                                totalAttempts: totalAttempts
-                            });
-                        }
-                    }, 1000);
-                }
-                else {
-                    // No more fallbacks
-                    console.log(`[Retry] All ${totalAttempts} session restoration attempts failed`);
-                    mainWindow?.webContents.send(`claude:restoration-status:${instanceId}`, {
-                        status: 'all-failed',
-                        totalAttempts: totalAttempts
-                    });
-                }
-            }
-            // Check for successful restoration - "Welcome to Claude Code" appears after successful start
-            // Strip ANSI escape codes to check plain text
-            const plainData = data.replace(/\x1b\[[0-9;]*m/g, '');
-            if (!hasUpdatedSuccessfulSession && plainData.includes('Welcome to Claude Code')) {
-                console.log(`[Retry] Session successfully restored with ID: ${sessionId}`);
-                // Update the session with the successful session ID
-                // Only update if we haven't already (prevent duplicate updates)
-                if (sessionId) {
-                    hasUpdatedSuccessfulSession = true;
-                    sessionService.updateSessionId(instanceId, sessionId);
-                }
-                mainWindow?.webContents.send(`claude:restoration-status:${instanceId}`, {
-                    status: 'success',
-                    sessionId: sessionId,
-                    attemptNumber: localAttemptIndex + 1
-                });
-                isRetrying = false; // Clear retry flag on success
-            }
-        });
-        // Setup exit handler
-        claudePty.onExit((event) => {
-            // Only send exit event if not killing for retry
-            if (!isKillingForRetry) {
-                mainWindow?.webContents.send(`claude:exit:${instanceId}`, event.exitCode);
-                instanceService.deleteInstance(instanceId);
-            }
-        });
-        return { success: true, pid: claudePty.pid };
-    }
-    catch (error) {
-        console.error(`Failed to start Claude with session ID ${sessionId}:`, error);
-        return {
-            success: false,
-            error: error instanceof Error ? error.message : String(error)
-        };
-    }
-}
 // Knowledge cache instances per workspace
 const knowledgeCaches = new Map();
 // Git service instances per workspace
@@ -200,45 +47,6 @@ const worktreeManagers = new Map();
 const snapshotServices = new Map();
 const isDev = process.env.NODE_ENV !== 'production';
 const nuxtURL = isDev ? 'http://localhost:3000' : `file://${join(__dirname, '../.output/public/index.html')}`;
-// Helper functions are now in ClaudeSessionService
-// Helper function to clean up all resources on reload/refresh
-function cleanupResourcesOnReload() {
-    console.log('Cleaning up resources on reload...');
-    // Don't kill Claude instances immediately - just clear the map
-    // The sessions will be restored after reload
-    if (instanceService.getInstanceCount() > 0) {
-        console.log(`Preserving ${instanceService.getInstanceCount()} Claude sessions for restoration...`);
-        // Mark all active sessions for auto-start
-        instanceService.getAllInstances().forEach((pty, instanceId) => {
-            const session = sessionService.get(instanceId);
-            if (session) {
-                session.shouldAutoStart = true;
-                console.log(`Marked session ${instanceId} for auto-start`);
-            }
-        });
-        // Save sessions to disk before clearing
-        sessionService.saveSessionsToDisk();
-        // Clear the instances map without killing processes
-        instanceService.clearInstances();
-    }
-    // Clean up terminal instances (these can't be restored easily)
-    if (terminals.size > 0) {
-        console.log(`Cleaning up ${terminals.size} terminal instances...`);
-        terminals.forEach((terminal, id) => {
-            console.log(`Killing terminal: ${id}`);
-            try {
-                terminal.kill();
-            }
-            catch (error) {
-                console.error(`Error killing terminal ${id}:`, error);
-            }
-        });
-        terminals.clear();
-    }
-    // No cleanup - sessions are kept indefinitely until explicitly deleted by user
-    console.log(`Preserved ${sessionService.size} recent Claude sessions for restoration`);
-    console.log('Resource cleanup completed');
-}
 function createWindow() {
     mainWindow = new BrowserWindow({
         width: 1600,
@@ -268,13 +76,13 @@ function createWindow() {
     });
     // Clean up resources when the renderer process navigates/reloads
     mainWindow.webContents.on('will-navigate', () => {
-        cleanupResourcesOnReload();
+        instanceService.cleanupResourcesOnReload();
     });
     // Also handle reload specifically
     mainWindow.webContents.on('before-input-event', (event, input) => {
         // Detect Cmd+R or Ctrl+R for reload
         if ((input.control || input.meta) && input.key === 'r' && input.type === 'keyDown') {
-            cleanupResourcesOnReload();
+            instanceService.cleanupResourcesOnReload();
         }
     });
     mainWindow.on('closed', () => {
@@ -311,430 +119,16 @@ app.on('window-all-closed', () => {
 });
 // Claude Process Management using PTY with multi-instance support
 ipcMain.handle('claude:start', async (event, instanceId, workingDirectory, instanceName, runConfig) => {
-    if (instanceService.hasInstance(instanceId)) {
-        return { success: false, error: 'Claude instance already running' };
-    }
-    // Check if we have a preserved session for this instance
-    const preservedSession = sessionService.get(instanceId);
-    // Always try to restore if session exists
-    const shouldRestore = !!preservedSession;
-    // If restoring, add a small delay to ensure clean state
-    if (shouldRestore) {
-        console.log(`Preparing to restore session for ${instanceId}...`);
-        // Small delay to ensure clean state
-        await new Promise(resolve => setTimeout(resolve, 1000));
-    }
-    try {
-        // Detect Claude installation
-        const claudeInfo = await ClaudeDetector.detectClaude(workingDirectory);
-        // Get the command configuration
-        const debugArgs = process.env.CLAUDE_DEBUG === 'true' ? ['--debug'] : [];
-        // Session management strategy:
-        // - Use --resume <sessionId> when restoring to continue the conversation
-        // - Use --session-id <uuid> for new sessions to have a specific ID
-        // - We need to capture the NEW session ID that Claude creates after resuming
-        let sessionUuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
-            const r = Math.random() * 16 | 0;
-            const v = c === 'x' ? r : (r & 0x3 | 0x8);
-            return v.toString(16);
-        });
-        // Build base arguments with fallback support
-        let baseArgs;
-        let sessionToRestore;
-        if (shouldRestore && preservedSession) {
-            // Get all possible session IDs (current + fallbacks)
-            const sessionOptions = sessionService.getSessionWithFallbacks(instanceId);
-            if (sessionOptions.fallbacks.length > 0) {
-                // We'll try the first one, and handle failures in the process output
-                sessionToRestore = sessionOptions.fallbacks[0];
-                baseArgs = ['--resume', sessionToRestore, ...debugArgs];
-                console.log(`Attempting to resume conversation with session ID ${sessionToRestore} for instance ${instanceId}`);
-                console.log(`Have ${sessionOptions.fallbacks.length - 1} fallback session IDs available if needed`);
-            }
-            else {
-                // No session IDs available, start fresh
-                baseArgs = ['--session-id', sessionUuid, ...debugArgs];
-                console.log(`No session IDs available for restoration, starting new session with ID ${sessionUuid}`);
-            }
-        }
-        else {
-            // Start a new session with a specific ID
-            baseArgs = ['--session-id', sessionUuid, ...debugArgs];
-            console.log(`Starting new session with ID ${sessionUuid} for instance ${instanceId}`);
-        }
-        let { command, args: commandArgs, useShell } = ClaudeDetector.getClaudeCommand(claudeInfo, baseArgs);
-        // Override with run config if provided
-        if (runConfig) {
-            if (runConfig.command) {
-                // If the command is just 'claude', use the detected path
-                command = runConfig.command === 'claude' ? command : runConfig.command;
-            }
-            if (runConfig.args && runConfig.args.length > 0) {
-                // When we have custom args, we need to rebuild the command
-                // Include resume or session-id based on whether we're restoring
-                let allArgs;
-                if (sessionToRestore) {
-                    allArgs = ['--resume', sessionToRestore, ...runConfig.args, ...debugArgs];
-                }
-                else {
-                    allArgs = ['--session-id', sessionUuid, ...runConfig.args, ...debugArgs];
-                }
-                const result = ClaudeDetector.getClaudeCommand(claudeInfo, allArgs);
-                command = result.command;
-                commandArgs = result.args;
-                useShell = result.useShell;
-            }
-        }
-        if (shouldRestore) {
-            console.log(`Attempting to restore Claude session for instance ${instanceId}`);
-        }
-        // Log settings file to verify it exists
-        const settingsPath = join(homedir(), '.claude', 'settings.json');
-        if (!existsSync(settingsPath)) {
-            console.warn('Claude settings file not found!');
-        }
-        // Get the user's default shell
-        const userShell = process.env.SHELL || '/bin/bash';
-        const claudePty = pty.spawn(command, commandArgs, {
-            name: 'xterm-color',
-            cols: 80,
-            rows: 30,
-            cwd: workingDirectory,
-            env: {
-                ...process.env,
-                FORCE_COLOR: '1',
-                TERM: 'xterm-256color',
-                HOME: process.env.HOME, // Ensure HOME is set so Claude can find ~/.claude/settings.json
-                USER: process.env.USER, // Ensure USER is set
-                SHELL: userShell, // Ensure SHELL is set
-                // Add instance-specific environment variables for hooks
-                CLAUDE_INSTANCE_ID: instanceId,
-                CLAUDE_INSTANCE_NAME: instanceName || `Claude-${instanceId.slice(7, 15)}`, // Use provided name or short ID
-                CLAUDE_IDE_INSTANCE: 'true'
-            }
-        });
-        // Store this instance
-        instanceService.setInstance(instanceId, claudePty);
-        // Store session data for potential restoration
-        // Store the initial session ID (either the one we're trying to resume or new UUID)
-        const initialSessionId = sessionToRestore || sessionUuid;
-        // If we're restoring, keep the existing session data with history
-        if (shouldRestore && preservedSession) {
-            // Update existing session, preserving history
-            preservedSession.lastActive = Date.now();
-            if (!sessionToRestore) {
-                // Starting fresh, but keep history
-                sessionService.updateSessionId(instanceId, sessionUuid);
-            }
-        }
-        else {
-            // Create new session
-            sessionService.set(instanceId, {
-                sessionId: initialSessionId,
-                instanceId,
-                workingDirectory,
-                instanceName,
-                runConfig,
-                lastActive: Date.now()
-            });
-        }
-        // Save to disk immediately
-        sessionService.saveSessionsToDisk();
-        if (shouldRestore) {
-            console.log(`Successfully restored Claude session for instance ${instanceId}`);
-        }
-        // After starting Claude, detect the new session file
-        // Skip immediate detection for restoration attempts, but schedule a later check
-        // to catch new sessions Claude creates after successful restoration
-        if (shouldRestore && !sessionToRestore) {
-            // Get the Claude project directory for this workspace
-            const projectDirName = workingDirectory.replace(/\//g, '-');
-            const claudeProjectDir = join(homedir(), '.claude', 'projects', projectDirName);
-            // Wait a bit for Claude to create the new session file
-            setTimeout(async () => {
-                try {
-                    if (existsSync(claudeProjectDir)) {
-                        // Get all .jsonl files sorted by creation time
-                        const files = readdirSync(claudeProjectDir)
-                            .filter(f => f.endsWith('.jsonl'))
-                            .map(f => {
-                            const fullPath = join(claudeProjectDir, f);
-                            const stats = statSync(fullPath);
-                            return { name: f.replace('.jsonl', ''), time: stats.mtimeMs };
-                        })
-                            .sort((a, b) => b.time - a.time);
-                        if (files.length > 0) {
-                            const newestSessionId = files[0].name;
-                            const session = sessionService.get(instanceId);
-                            if (session && newestSessionId !== session.sessionId) {
-                                console.log(`[Claude Session] Updated session ID for ${instanceId}: ${newestSessionId} (was: ${session.sessionId})`);
-                                sessionService.updateSessionId(instanceId, newestSessionId);
-                                // Send log to renderer
-                                mainWindow?.webContents.send('log', `Session updated to: ${newestSessionId}`);
-                            }
-                        }
-                    }
-                }
-                catch (error) {
-                    console.error('Failed to detect new session ID:', error);
-                }
-            }, 3000); // Wait 3 seconds for Claude to create the session file
-        }
-        // Track session restoration attempts
-        let sessionRestorationFailed = false;
-        let currentAttemptIndex = 0;
-        const sessionOptions = sessionService.getSessionWithFallbacks(instanceId);
-        let isRetrying = false;
-        let isKillingForRetry = false; // Flag to prevent exit handler during retry
-        let hasUpdatedSuccessfulSession = false; // Prevent duplicate session updates
-        // Handle output from Claude
-        claudePty.onData((data) => {
-            // Detect session restoration failure and automatically retry with fallbacks
-            if (!isRetrying && data.includes('No conversation found with session ID:')) {
-                sessionRestorationFailed = true;
-                const failedSessionId = sessionOptions.fallbacks[currentAttemptIndex] || 'unknown';
-                console.log(`[Claude Session] Session restoration failed for ${instanceId} with session ID: ${failedSessionId}`);
-                // Send restoration status to renderer
-                mainWindow?.webContents.send(`claude:restoration-status:${instanceId}`, {
-                    status: 'failed',
-                    sessionId: failedSessionId,
-                    attemptNumber: currentAttemptIndex + 1,
-                    totalAttempts: sessionOptions.fallbacks.length
-                });
-                // Check if we have more fallbacks to try
-                currentAttemptIndex++;
-                if (currentAttemptIndex < sessionOptions.fallbacks.length) {
-                    const nextSessionId = sessionOptions.fallbacks[currentAttemptIndex];
-                    console.log(`[Claude Session] Automatically retrying with fallback session ID ${currentAttemptIndex + 1}/${sessionOptions.fallbacks.length}: ${nextSessionId}`);
-                    // Send retry status to renderer
-                    mainWindow?.webContents.send(`claude:restoration-status:${instanceId}`, {
-                        status: 'retrying',
-                        sessionId: nextSessionId,
-                        attemptNumber: currentAttemptIndex + 1,
-                        totalAttempts: sessionOptions.fallbacks.length
-                    });
-                    // Kill current process and restart with fallback
-                    isRetrying = true;
-                    isKillingForRetry = true; // Set flag to prevent exit handler
-                    // Kill the current Claude process
-                    try {
-                        claudePty.kill();
-                    }
-                    catch (error) {
-                        console.error('Error killing Claude process for retry:', error);
-                    }
-                    // Remove from instances map
-                    instanceService.deleteInstance(instanceId);
-                    // Wait a bit before retrying
-                    setTimeout(async () => {
-                        // Restart Claude with the fallback session ID
-                        const retryResult = await startClaudeWithSessionId(instanceId, workingDirectory, instanceName, runConfig, nextSessionId, currentAttemptIndex, sessionOptions.fallbacks.length);
-                        if (!retryResult.success) {
-                            console.error(`Failed to retry with fallback session ID: ${retryResult.error}`);
-                            mainWindow?.webContents.send(`claude:restoration-status:${instanceId}`, {
-                                status: 'all-failed',
-                                error: retryResult.error
-                            });
-                        }
-                    }, 1000);
-                }
-                else {
-                    // No more fallbacks, session restoration completely failed
-                    console.log(`[Claude Session] All ${sessionOptions.fallbacks.length} session restoration attempts failed for ${instanceId}`);
-                    mainWindow?.webContents.send(`claude:restoration-status:${instanceId}`, {
-                        status: 'all-failed',
-                        totalAttempts: sessionOptions.fallbacks.length
-                    });
-                }
-            }
-            // Detect successful restoration - "Welcome to Claude Code" appears after successful start
-            // Strip ANSI escape codes to check plain text
-            const plainData = data.replace(/\x1b\[[0-9;]*m/g, '');
-            if (!hasUpdatedSuccessfulSession && plainData.includes('Welcome to Claude Code')) {
-                const successfulSessionId = sessionOptions.fallbacks[currentAttemptIndex];
-                console.log(`[Claude Session] Session successfully restored for ${instanceId} with session ID: ${successfulSessionId}`);
-                // Update the session with the successful session ID so next pause/continue uses the right one
-                // Only update if we haven't already (prevent duplicate updates)
-                hasUpdatedSuccessfulSession = true;
-                sessionService.updateSessionId(instanceId, successfulSessionId);
-                mainWindow?.webContents.send(`claude:restoration-status:${instanceId}`, {
-                    status: 'success',
-                    sessionId: successfulSessionId,
-                    attemptNumber: currentAttemptIndex + 1
-                });
-                isRetrying = false; // Clear retry flag on success
-                // After successful restoration, schedule a check for new session files
-                // Claude might create a new session when you continue working
-                // 
-                // NOTE: Tab safety limitations
-                // This is still not 100% tab-safe because:
-                // - If Tab B creates a new session AFTER Tab A's restoration time (within that 10-second window), 
-                //   Tab A might still pick it up
-                // - There's no way to definitively link a Claude session file to a specific tab instance
-                //
-                // A truly tab-safe solution would require Claude to provide instance-specific session IDs or some way 
-                // to track which session belongs to which terminal instance. For now, this time-based filtering is 
-                // the best we can do to minimize cross-tab interference.
-                const restorationTime = Date.now();
-                setTimeout(async () => {
-                    try {
-                        const projectDirName = workingDirectory.replace(/\//g, '-');
-                        const claudeProjectDir = join(homedir(), '.claude', 'projects', projectDirName);
-                        if (existsSync(claudeProjectDir)) {
-                            const files = readdirSync(claudeProjectDir)
-                                .filter(f => f.endsWith('.jsonl'))
-                                .map(f => {
-                                const fullPath = join(claudeProjectDir, f);
-                                const stats = statSync(fullPath);
-                                return {
-                                    name: f.replace('.jsonl', ''),
-                                    time: stats.mtimeMs,
-                                    createdAfterRestore: stats.mtimeMs > restorationTime
-                                };
-                            })
-                                .filter(f => f.createdAfterRestore) // Only consider files created after restoration
-                                .sort((a, b) => b.time - a.time);
-                            if (files.length > 0) {
-                                const newestSessionId = files[0].name;
-                                const currentSession = sessionService.get(instanceId);
-                                // Only update if we find a session created after this instance's restoration
-                                if (currentSession && newestSessionId !== currentSession.sessionId) {
-                                    console.log(`[Claude Session] Detected new session file after restoration for ${instanceId}: ${newestSessionId}`);
-                                    sessionService.updateSessionId(instanceId, newestSessionId);
-                                }
-                            }
-                        }
-                    }
-                    catch (error) {
-                        console.error('Failed to check for new session after restoration:', error);
-                    }
-                }, 10000); // Check after 10 seconds to give Claude time to create new session
-            }
-            // Update last active time on any output
-            const session = sessionService.get(instanceId);
-            if (session) {
-                session.lastActive = Date.now();
-            }
-            // Send data with instance ID
-            mainWindow?.webContents.send(`claude:output:${instanceId}`, data);
-        });
-        // Handle exit
-        claudePty.onExit(({ exitCode, signal }) => {
-            // Don't send exit event if we're killing for retry
-            if (!isKillingForRetry) {
-                mainWindow?.webContents.send(`claude:exit:${instanceId}`, exitCode);
-                instanceService.deleteInstance(instanceId);
-                // Clear auto-start flag when session exits normally
-                const session = sessionService.get(instanceId);
-                if (session) {
-                    session.shouldAutoStart = false;
-                    sessionService.saveSessionsToDisk();
-                }
-            }
-        });
-        return {
-            success: true,
-            pid: claudePty.pid,
-            claudeInfo: {
-                path: claudeInfo.path,
-                version: claudeInfo.version,
-                source: claudeInfo.source
-            },
-            restored: shouldRestore
-        };
-    }
-    catch (error) {
-        console.error(`Failed to start Claude for ${instanceId}:`, error);
-        // If restoration failed, try again without --continue
-        if (shouldRestore) {
-            console.log(`Session restoration failed for ${instanceId}, starting fresh session...`);
-            try {
-                // Re-detect Claude installation for fallback
-                const claudeInfo = await ClaudeDetector.detectClaude(workingDirectory);
-                // Get fresh command without --continue
-                const debugArgs = process.env.CLAUDE_DEBUG === 'true' ? ['--debug'] : [];
-                let { command, args: commandArgs } = ClaudeDetector.getClaudeCommand(claudeInfo, debugArgs);
-                // Apply run config if provided
-                if (runConfig && runConfig.args && runConfig.args.length > 0) {
-                    const allArgs = [...runConfig.args, ...debugArgs];
-                    const result = ClaudeDetector.getClaudeCommand(claudeInfo, allArgs);
-                    command = result.command;
-                    commandArgs = result.args;
-                }
-                // Start fresh Claude instance
-                const claudePty = pty.spawn(command, commandArgs, {
-                    name: 'xterm-color',
-                    cols: 80,
-                    rows: 30,
-                    cwd: workingDirectory,
-                    env: {
-                        ...process.env,
-                        FORCE_COLOR: '1',
-                        TERM: 'xterm-256color',
-                        HOME: process.env.HOME,
-                        USER: process.env.USER,
-                        SHELL: process.env.SHELL || '/bin/bash',
-                        CLAUDE_INSTANCE_ID: instanceId,
-                        CLAUDE_INSTANCE_NAME: instanceName || `Claude-${instanceId.slice(7, 15)}`,
-                        CLAUDE_IDE_INSTANCE: 'true'
-                    }
-                });
-                // Store instance and session data
-                instanceService.setInstance(instanceId, claudePty);
-                sessionService.set(instanceId, {
-                    instanceId,
-                    workingDirectory,
-                    instanceName,
-                    runConfig,
-                    lastActive: Date.now()
-                });
-                // Set up handlers
-                claudePty.onData((data) => {
-                    const session = sessionService.get(instanceId);
-                    if (session) {
-                        session.lastActive = Date.now();
-                    }
-                    mainWindow?.webContents.send(`claude:output:${instanceId}`, data);
-                });
-                claudePty.onExit(({ exitCode }) => {
-                    mainWindow?.webContents.send(`claude:exit:${instanceId}`, exitCode);
-                    instanceService.deleteInstance(instanceId);
-                });
-                console.log(`Started fresh Claude session for ${instanceId} after restoration failed`);
-                return {
-                    success: true,
-                    pid: claudePty.pid,
-                    claudeInfo: {
-                        path: claudeInfo.path,
-                        version: claudeInfo.version,
-                        source: claudeInfo.source
-                    },
-                    restored: false,
-                    restorationFailed: true
-                };
-            }
-            catch (fallbackError) {
-                console.error(`Fallback to fresh session also failed for ${instanceId}:`, fallbackError);
-                return { success: false, error: fallbackError instanceof Error ? fallbackError.message : String(fallbackError) };
-            }
-        }
-        return { success: false, error: error instanceof Error ? error.message : String(error) };
-    }
+    return instanceService.startClaude(instanceId, workingDirectory, instanceName, runConfig);
 });
 ipcMain.handle('claude:send', async (event, instanceId, command) => {
-    const claudePty = instanceService.getInstance(instanceId);
-    if (!claudePty) {
-        return { success: false, error: `Claude instance is not running. Please start a Claude instance in the terminal first.` };
-    }
-    try {
-        // Write raw data to PTY (xterm.js will handle line endings)
-        claudePty.write(command);
-        return { success: true };
-    }
-    catch (error) {
-        console.error(`Failed to send command to Claude PTY ${instanceId}:`, error);
-        return { success: false, error: error instanceof Error ? error.message : String(error) };
-    }
+    return instanceService.sendCommand(instanceId, command);
+});
+ipcMain.handle('claude:stop', async (event, instanceId) => {
+    return instanceService.stopInstance(instanceId);
+});
+ipcMain.handle('claude:resize', async (event, instanceId, cols, rows) => {
+    return instanceService.resizeTerminal(instanceId, cols, rows);
 });
 // Get preserved sessions that should auto-start
 ipcMain.handle('claude:getPreservedSessions', async () => {
@@ -752,100 +146,14 @@ ipcMain.handle('claude:clearAutoStart', async (event, instanceId) => {
     }
     return { success: true };
 });
-// Pause Claude instance (keeps session data)
-ipcMain.handle('claude:stop', async (event, instanceId) => {
-    const claudePty = instanceService.getInstance(instanceId);
-    if (claudePty) {
-        try {
-            console.log(`Stopping Claude instance ${instanceId} (PID: ${claudePty.pid})...`);
-            // Send SIGINT (Ctrl+C) for graceful shutdown
-            // This should trigger Claude to save the session
-            claudePty.kill('SIGINT');
-            console.log(`Sent SIGINT to process ${claudePty.pid}`);
-            // Wait for graceful shutdown and session save
-            // Claude likely auto-saves on exit, so give it time
-            await new Promise(resolve => setTimeout(resolve, 2000));
-            // Check if process is still alive and send SIGTERM if needed
-            try {
-                process.kill(claudePty.pid, 0); // Check if process exists
-                console.log(`Process ${claudePty.pid} still alive, sending SIGTERM...`);
-                claudePty.kill('SIGTERM');
-                await new Promise(resolve => setTimeout(resolve, 500));
-            }
-            catch (e) {
-                // Process already dead, which is good
-                console.log(`Process ${claudePty.pid} terminated gracefully`);
-            }
-            // Final force kill if still running
-            try {
-                process.kill(claudePty.pid, 0); // Check if process exists
-                console.log(`Process ${claudePty.pid} still alive, forcing SIGKILL...`);
-                claudePty.kill('SIGKILL');
-            }
-            catch (e) {
-                // Process already dead
-            }
-        }
-        catch (error) {
-            console.error(`Error stopping Claude instance ${instanceId}:`, error);
-        }
-        instanceService.deleteInstance(instanceId);
-        // Ensure session data is saved
-        const session = sessionService.get(instanceId);
-        if (session) {
-            session.lastActive = Date.now();
-            sessionService.saveSessionsToDisk();
-            console.log(`Session data saved for instance ${instanceId}`);
-        }
-        return { success: true };
-    }
-    return { success: false, error: `No Claude PTY running for instance ${instanceId}` };
-});
 // Stop and delete Claude session permanently
 ipcMain.handle('claude:deleteSession', async (event, instanceId) => {
     console.log(`Deleting session for instance ${instanceId}...`);
-    // Stop the process if running (with graceful shutdown)
-    const claudePty = instanceService.getInstance(instanceId);
-    if (claudePty) {
-        try {
-            // Send SIGINT for graceful shutdown
-            claudePty.kill('SIGINT');
-            await new Promise(resolve => setTimeout(resolve, 1000));
-            // Force kill if needed
-            try {
-                process.kill(claudePty.pid, 0);
-                claudePty.kill('SIGKILL');
-            }
-            catch (e) {
-                // Process already dead
-            }
-        }
-        catch (error) {
-            console.error(`Error stopping Claude instance ${instanceId}:`, error);
-        }
-        instanceService.deleteInstance(instanceId);
-    }
-    // Remove from session storage
-    sessionService.delete(instanceId);
-    // Delete session file from disk
-    try {
-        const workspacePath = store.get('workspacePath') || process.cwd();
-        const sessionsDir = join(workspacePath, '.claude', 'sessions');
-        const sessionFile = join(sessionsDir, `${instanceId}.json`);
-        if (existsSync(sessionFile)) {
-            unlinkSync(sessionFile);
-        }
-        // Update manifest
-        sessionService.saveSessionsToDisk();
-        return { success: true };
-    }
-    catch (error) {
-        console.error(`Failed to delete session ${instanceId}:`, error);
-        return { success: false, error: error instanceof Error ? error.message : String(error) };
-    }
-});
-ipcMain.handle('claude:resize', async (event, instanceId, cols, rows) => {
-    return instanceService.resizeTerminal(instanceId, cols, rows);
+    // Stop the instance if running
+    await instanceService.stopInstance(instanceId);
+    // Delete session and its file
+    const deleted = sessionService.deleteSessionAndFile(instanceId);
+    return { success: deleted };
 });
 // Get home directory
 ipcMain.handle('getHomeDir', () => {
@@ -2106,7 +1414,6 @@ Remember: Return ONLY the complete code for the file. No explanations. No markdo
             const response = query({
                 prompt: userPrompt,
                 options: {
-                    abortController,
                     model: 'claude-sonnet-4-20250514', // Fast model for code generation
                     maxTurns: 1,
                     allowedTools: [],
